+++ handlers/doc.go
@@ -23,7 +23,7 @@ import (
 
 // PackageVersion returns the version of the version package.
 func PackageVersion() version.Version {
-	return version.New(2, 1, 1)
+	return version.New(2, 2, 0)
 }
 
 // EOF