+++ rest/context.go
@@ -22,20 +22,38 @@ import (
 // contextKey is used to address data inside a context.
 type contextKey int
 
-// jobKey addresses the worksheet inside the context.
-const jobKey contextKey = 0
+const (
+	// envKey addresses the environment inside the context.
+	envKey contextKey = 0
+
+	// jobKey addresses the job inside the context.
+	jobKey contextKey = 1
+)
 
 //--------------------
 // CONTEXT
 //--------------------
 
-// newContext creates a context containing the passed job.
-func newContext(job Job) context.Context {
-	return context.WithValue(context.Background(), jobKey, job)
+// newEnvironmentContext creates a context based on the passed one
+// and containing the passed environment.
+func newEnvironmentContext(ctx context.Context, env Environment) context.Context {
+	return context.WithValue(ctx, envKey, env)
 }
 
-// FromContext retrieves the job out of a context.
-func FromContext(ctx context.Context) (Job, bool) {
+// newJobContext creates a context based on the passed one
+// and containing the passed job.
+func newJobContext(ctx context.Context, job Job) context.Context {
+	return context.WithValue(ctx, jobKey, job)
+}
+
+// EnvironmentFromContext retrieves the environment out of a context.
+func EnvironmentFromContext(ctx context.Context) (Environment, bool) {
+	env, ok := ctx.Value(envKey).(Environment)
+	return env, ok
+}
+
+// JobFromContext retrieves the job out of a context.
+func JobFromContext(ctx context.Context) (Job, bool) {
 	job, ok := ctx.Value(jobKey).(Job)
 	return job, ok
 }